Refactor fare calculation to use integer values
Changed price and money handling from float to int for improved accuracy and consistency. Updated related function signatures and logic in io, main, monCalc, and tramManager. Also restructured the change struct and improved input validation and user interaction flow in main.cpp.

+++ src/tramManager.h
@@ -6,7 +6,7 @@ using namespace std;
 
 struct TramLine {
     int tram_number;
-    float price_per_station;
+    int price_per_station;
     vector<string> stations;
 };
